web_api_service/document/nlp_module/requirements.txt
tensorflow_text==2.3.0
tensorflow==2.3.0
scikit-learn==0.23.2
scipy==1.4.1
numpy==1.18.5
pandas==1.1.3
tensorflow_hub==0.8.0
joblib
natasha
PyMySQL>=0.10.1
pika==1.1.0
requests
google-cloud-vision>=2.0.0
Pillow>=7.2.0
pytesseract
